move struct back
