+++ src/EnergyPlus/ExteriorEnergyUse.hh
@@ -53,10 +53,12 @@
 
 // EnergyPlus Headers
 #include <EnergyPlus/DataGlobals.hh>
-#include <EnergyPlus/Data/EnergyPlusData.hh>
 #include <EnergyPlus/EnergyPlus.hh>
 
 namespace EnergyPlus {
+
+struct EnergyPlusData;
+struct ExteriorEnergyUseData;
 
 namespace ExteriorEnergyUse {
 
@@ -78,6 +80,23 @@ namespace ExteriorEnergyUse {
 
     // Types
 
+    struct ExteriorEquipmentUsage
+    {
+        // Members
+        std::string Name; // Descriptive name -- will show on reporting
+        int FuelType;
+        int SchedPtr;       // Can be scheduled
+        Real64 DesignLevel; // Design Consumption (Watts, except for Water Equipment)
+        Real64 Power;       // Power = DesignLevel * ScheduleValue
+        Real64 CurrentUse;  // Use for this time step
+        bool ManageDemand;  // Flag to indicate whether to use demand limiting
+        Real64 DemandLimit; // Demand limit set by demand manager [W]
+
+                            // Default Constructor
+        ExteriorEquipmentUsage() : FuelType(0), SchedPtr(0), DesignLevel(0.0), Power(0.0), CurrentUse(0.0), ManageDemand(false), DemandLimit(0.0)
+        {
+        }
+    };
     // Object Data
 
     // Functions
